Set metadata to route health events to correct index and set source
Currently the health events are being routed to index=csls_dev which
is not ideal.

For now my intention is to route them to cyber_services_prod|test
and set the source field to "health_monitoring_event".

We can then create an eventtype which looks for the data in both
places so we don't lose the old data.