Fix multiple datapoints from get-metric-statistics
We were setting the time period to 28 days ago to now and setting
the period to exactly 28 days in seconds but this was occasionally
returning multiple datapoints.

It's not entirely straight forward to aggregate multiple datapoints
since the stat could be Max/Min/Avg... so you'd have to be sensitive
to that in how you aggregated.

It's easier to force it to be a single datapoint by making the
period longer than the elapsed time.