Refactor how metric region is calculated for forwarder lambdas
In the generate_metric_alarms script the metrics are enriched
with the region meaning that from that point onwards we can
refer to metric.Region.

The clouwatch metric/alarm forwarder lambdas are deployed per
region and so the session region for the lambda matches the metric
being forwarded. In this case the metric.Region property is not
set and the code defaults to the session region.