Handle alarm config change better to make Slack quieter.
When alarms are created or changed they go into INSUFFICIENT_DATA.
When they resolve if they resolve to healthy we don't need to see
those events in Slack.

We would want to see them in Slack if they resolve to UNHEALTHY
and we would want to see them in Splunk regardless.